Fix download queue bugs, add stop confirmation, clean up codebase
Bug fixes:
- Fix pause/resume: queue.clear() prevents task drain while paused;
  resumeQueue() re-adds all waiting tasks using stored currentMirrors/currentOptions;
  race condition guard re-adds aborted tasks if resume ran before abort settled
- Fix stop download: clearQueue() now destroys active HTTP requests before clearing;
  httpDownloader destroys response stream and writer on abort to stop file writes immediately

UI improvements:
- Add inline two-step stop confirmation (Stop -> confirm/cancel buttons)
- Show status message after queue completes (success/failed count) or is cancelled
- Disable pause button while confirming stop

Refactoring and cleanup:
- Modularize downloadService.ts into download/types.ts, fileUtils.ts, httpDownloader.ts
- Extract download settings persistence into useDownloadSettings composable
- Reuse validateDownloadPath from fileUtils in api.ts
- Consolidate .text-success/.text-error into main.css; remove duplicate Google Fonts imports
- Remove debug console.log statements throughout main process and components
- Add .eslintcache and .cursorindexingignore to .gitignore

Made-with: Cursor
