+++ src/renderer/src/i18n/index.ts
@@ -12,4 +12,4 @@ const i18n = createI18n({
   }
 })
 
-export default i18n 
+export default i18n
