+++ src/renderer/src/i18n/languageProperties.ts
@@ -1,13 +1,13 @@
 export const languageNames = {
   en: 'English',
-  vi: 'Tiếng Việt',
+  vi: 'Tiếng Việt'
   // Add new languages here
 } as const
 
 export const languageFlags = {
   en: 'gb',
-  vi: 'vn',
+  vi: 'vn'
   // Add new language flags here
 } as const
 
-export type LanguageCode = keyof typeof languageNames 
+export type LanguageCode = keyof typeof languageNames
