+++ src/services/downloadService.ts
@@ -1,62 +1,20 @@
 import { EventEmitter } from 'events'
-// import { default as PQueue } from 'p-queue'
-import path from 'path'
+import { BeatmapMirror, DefaultBeatmapMirrors } from '../config/beatmapMirrors'
+import { DownloadTask, DownloadOptions, DownloadEvent } from './download/types'
+import {
+  getDefaultDownloadPath,
+  validateDownloadPath,
+  validateBackupFile,
+  getExistingBeatmapsetIds
+} from './download/fileUtils'
+import { downloadFile, MirrorHealth } from './download/httpDownloader'
 import fs from 'fs'
-import https from 'https'
-import http from 'http'
-import { URL } from 'url'
-import { BeatmapMirror } from '../config/beatmapMirrors'
-import { DefaultBeatmapMirrors } from '../config/beatmapMirrors'
-import os from 'os'
-import { execSync } from 'child_process'
-import { getOsuStablePath } from './settingsStore'
-import { realmService } from './realmService'
+
+export type { DownloadTask, DownloadOptions }
+export { DownloadEvent }
 
 // eslint-disable-next-line @typescript-eslint/no-require-imports
 const PQueue = require('p-queue').default
-console.log('PQueue in downloadService:', PQueue)
-console.log('PQueue type:', typeof PQueue)
-console.log('PQueue prototype:', PQueue.prototype)
-
-// Types
-export interface DownloadTask {
-  id: string
-  beatmapsetId: string
-  mirror: BeatmapMirror
-  noVideo: boolean
-  status: 'waiting' | 'downloading' | 'completed' | 'error'
-  progress: number
-  speed: number
-  remainingTime: number
-  error?: string
-  // Directory chosen by user (or default OS Downloads)
-  downloadPath?: string
-  // Final file name and full path, set once headers are known / completed
-  fileName?: string
-  filePath?: string
-  request?: ReturnType<typeof https.get>
-}
-
-export interface DownloadOptions {
-  threadCount: number
-  sources: string[]
-  noVideo: boolean
-  downloadPath?: string
-  removeFromStable: boolean
-  removeFromLazer: boolean
-}
-
-// Events
-export enum DownloadEvent {
-  TASK_ADDED = 'taskAdded',
-  TASK_UPDATED = 'taskUpdated',
-  TASK_COMPLETED = 'taskCompleted',
-  TASK_ERROR = 'taskError',
-  QUEUE_PAUSED = 'queuePaused',
-  QUEUE_RESUMED = 'queueResumed',
-  QUEUE_CLEARED = 'queueCleared',
-  QUEUE_COMPLETED = 'queueCompleted'
-}
 
 class DownloadService extends EventEmitter {
   private static instance: DownloadService
@@ -65,8 +23,11 @@ class DownloadService extends EventEmitter {
   private isPaused: boolean
   private cooldownPeriod: number
   private cooldownTimeout?: NodeJS.Timeout
-  private mirrorHealth: Map<string, { success: number; failure: number; avgResponseTime: number }>
+  private mirrorHealth: Map<string, MirrorHealth>
   private queueStartTime: number | null
+  /** Stored so resumeQueue can re-add waiting tasks without needing closure args */
+  private currentMirrors: BeatmapMirror[] = []
+  private currentOptions: DownloadOptions | null = null
 
   private constructor() {
     super()
@@ -85,178 +46,33 @@ class DownloadService extends EventEmitter {
     return DownloadService.instance
   }
 
-  private getDefaultDownloadPath(): string {
-    const platform = process.platform
-    const homeDir = os.homedir()
-
-    switch (platform) {
-      case 'win32':
-        try {
-          // Read registry to get Downloads folder location
-          const command =
-            'reg query "HKCU\\Software\\Microsoft\\Windows\\CurrentVersion\\Explorer\\User Shell Folders" /v "{374DE290-123F-4565-9164-39C4925E467B}"'
-          const output = execSync(command, { encoding: 'utf-8' })
-          const match = output.match(/REG_EXPAND_SZ\s+(.+)$/)
-
-          if (match) {
-            // Expand environment variables in the path
-            const expandedPath = match[1].replace(/%([^%]+)%/g, (_, n) => process.env[n] || '')
-            return expandedPath
-          }
-        } catch (error) {
-          console.warn('Failed to read registry for Downloads folder:', error)
-        }
-        // Fallback to default Downloads folder
-        return path.join(homeDir, 'Downloads')
-      case 'darwin':
-        return path.join(homeDir, 'Downloads')
-      case 'linux':
-        return path.join(homeDir, 'Downloads')
-      default:
-        return path.join(homeDir, 'Downloads')
-    }
-  }
-
-  private async getExistingBeatmapsetIds(options: DownloadOptions): Promise<Set<number>> {
-    const existingIds = new Set<number>()
-
-    if (options.removeFromStable) {
-      try {
-        const osuStablePath = getOsuStablePath()
-        if (osuStablePath) {
-          const songsPath = path.join(osuStablePath, 'Songs')
-          if (fs.existsSync(songsPath)) {
-            const folders = fs.readdirSync(songsPath)
-            for (const folder of folders) {
-              const match = folder.match(/^(\d+)\s/)
-              if (match) {
-                const id = parseInt(match[1])
-                if (!isNaN(id)) {
-                  existingIds.add(id)
-                }
-              }
-            }
-            console.log('Found', existingIds.size, 'existing stable beatmapset IDs')
-          }
-        }
-      } catch (error) {
-        console.warn('Failed to get stable beatmapset IDs:', error)
-      }
-    }
-
-    if (options.removeFromLazer) {
-      try {
-        const lazerIds = await realmService.getBeatmapsetIds()
-        lazerIds.forEach((id) => existingIds.add(id))
-        console.log('Found', lazerIds.length, 'existing lazer beatmapset IDs')
-      } catch (error) {
-        console.warn('Failed to get lazer beatmapset IDs:', error)
-      }
-    }
-
-    return existingIds
-  }
-
-  private validateBackupFile(content: string): void {
-    // Check if file has header
-    if (!content.startsWith('# Beatmap Backup File')) {
-      throw new Error('Invalid backup file format: Missing header')
-    }
-
-    // Check if file has required metadata
-    const requiredMetadata = [
-      '# Format: One beatmapset ID per line',
-      '# Created:',
-      '# Total beatmaps:',
-      '# Source:'
-    ]
-
-    for (const metadata of requiredMetadata) {
-      if (!content.includes(metadata)) {
-        throw new Error(`Invalid backup file format: Missing ${metadata}`)
-      }
-    }
-
-    // Get beatmapset IDs (skip header)
-    const lines = content.split('\n')
-    const ids = lines.filter((line) => line.trim() && !line.startsWith('#'))
-
-    if (ids.length === 0) {
-      throw new Error('Invalid backup file: No beatmapset IDs found')
-    }
-
-    // Validate each ID
-    for (const id of ids) {
-      if (!/^\d+$/.test(id.trim())) {
-        throw new Error(`Invalid beatmapset ID: ${id}`)
-      }
-    }
-  }
-
-  private async validateDownloadPath(downloadPath: string): Promise<void> {
-    try {
-      // Check if path exists
-      if (!fs.existsSync(downloadPath)) {
-        throw new Error('Download path does not exist')
-      }
-
-      // Check if path is a directory
-      const stats = await fs.promises.stat(downloadPath)
-      if (!stats.isDirectory()) {
-        throw new Error('Download path is not a directory')
-      }
-
-      // Check write permission
-      try {
-        await fs.promises.access(downloadPath, fs.constants.W_OK)
-      } catch {
-        throw new Error('No write permission in download path')
-      }
-    } catch (error) {
-      if (error instanceof Error) {
-        throw new Error(`Download path validation failed: ${error.message}`)
-      }
-      throw error
-    }
-  }
-
   public async startDownload(filePath: string, options: DownloadOptions): Promise<void> {
     try {
       this.queueStartTime = Date.now()
-      // Read and validate backup file
+
       const content = await fs.promises.readFile(filePath, 'utf-8')
-      this.validateBackupFile(content)
+      validateBackupFile(content)
       const beatmapsetIds = content
         .split('\n')
         .filter((line) => line.trim() && !line.startsWith('#'))
         .map((id) => id.trim())
 
-      // Get existing beatmapset IDs if ignore options are enabled
       let existingIds = new Set<number>()
       if (options.removeFromStable || options.removeFromLazer) {
-        existingIds = await this.getExistingBeatmapsetIds(options)
-        console.log('Found', existingIds.size, 'existing beatmapset IDs to ignore')
+        existingIds = await getExistingBeatmapsetIds(options)
       }
 
-      // Filter out existing beatmaps
       const filteredIds = beatmapsetIds.filter((id) => {
         const numericId = parseInt(id)
         return !isNaN(numericId) && !existingIds.has(numericId)
       })
 
-      console.log('Original beatmapset IDs:', beatmapsetIds.length)
-      console.log('After filtering existing:', filteredIds.length)
-      console.log('Ignored existing:', beatmapsetIds.length - filteredIds.length)
-
       if (filteredIds.length === 0) {
-        console.log('All beatmaps already exist, no download needed')
         return
       }
 
-      // Update queue concurrency
       this.queue.concurrency = options.threadCount
 
-      // Get available mirrors
       const availableMirrors = DefaultBeatmapMirrors.filter((mirror) =>
         options.sources.includes(mirror.name)
       )
@@ -265,13 +81,12 @@ class DownloadService extends EventEmitter {
         throw new Error('No available mirrors selected')
       }
 
-      // Use default download path if not specified
-      const downloadPath = options.downloadPath || this.getDefaultDownloadPath()
+      const dlPath = options.downloadPath || getDefaultDownloadPath()
+      await validateDownloadPath(dlPath)
 
-      // Validate download path
-      await this.validateDownloadPath(downloadPath)
+      this.currentMirrors = availableMirrors
+      this.currentOptions = options
 
-      // Add tasks to queue
       for (const beatmapsetId of filteredIds) {
         const task: DownloadTask = {
           id: `${beatmapsetId}-${Date.now()}`,
@@ -282,17 +97,14 @@ class DownloadService extends EventEmitter {
           progress: 0,
           speed: 0,
           remainingTime: 0,
-          downloadPath
+          downloadPath: dlPath
         }
 
         this.tasks.set(task.id, task)
         this.emit(DownloadEvent.TASK_ADDED, task)
-
         this.queue.add(() => this.downloadTask(task, availableMirrors, options))
       }
 
-      // When the queue becomes idle (no pending/active tasks), verify completion
-      // This covers timing where individual task callbacks resolve before PQueue updates counters
       this.queue.onIdle().then(() => {
         this.checkQueueCompletion()
       })
@@ -303,7 +115,6 @@ class DownloadService extends EventEmitter {
   }
 
   private checkQueueCompletion(): void {
-    // Not completed if there are active tasks, pending queue, or cooldown scheduled
     const hasActive = Array.from(this.tasks.values()).some(
       (t) => t.status !== 'completed' && t.status !== 'error'
     )
@@ -315,24 +126,23 @@ class DownloadService extends EventEmitter {
     const success = Array.from(this.tasks.values()).filter((t) => t.status === 'completed').length
     const failed = Array.from(this.tasks.values()).filter((t) => t.status === 'error').length
     const anyTask = this.tasks.values().next().value as DownloadTask | undefined
-    const downloadPath = anyTask?.downloadPath ?? null
+    const dlPath = anyTask?.downloadPath ?? null
     const durationMs = this.queueStartTime ? Date.now() - this.queueStartTime : 0
 
     this.emit(DownloadEvent.QUEUE_COMPLETED, {
       total,
       success,
       failed,
-      downloadPath,
+      downloadPath: dlPath,
       durationMs
     })
 
-    // Auto clear queue after short delay to let UI display completion
     setTimeout(() => this.clearQueue(), 2000)
   }
 
   private async downloadTask(
     task: DownloadTask,
-    availableMirrors: BeatmapMirror[],
+    availableMirrors: typeof DefaultBeatmapMirrors,
     options: DownloadOptions
   ): Promise<void> {
     if (this.isPaused) {
@@ -344,187 +154,27 @@ class DownloadService extends EventEmitter {
     task.status = 'downloading'
     this.emit(DownloadEvent.TASK_UPDATED, task)
 
+    const taskDownloadPath = task.downloadPath || options.downloadPath || getDefaultDownloadPath()
+
     try {
-      const downloadUrl = task.mirror.getDownloadUrl(task.beatmapsetId, task.noVideo)
-      // Prefer validated path stored on task, then provided option, then OS default
-      let downloadPath = task.downloadPath || options.downloadPath || this.getDefaultDownloadPath()
-      // If a drive root like "F:\\" is passed, default to a safe subfolder
-      const resolvedForCheck = path.resolve(downloadPath)
-      if (resolvedForCheck === path.parse(resolvedForCheck).root) {
-        downloadPath = path.join(downloadPath, 'osu-beatmaps')
-      }
-
-      // Create download directory if it doesn't exist (avoid creating drive root)
-      const resolvedPath = path.resolve(downloadPath)
-      const isRoot = resolvedPath === path.parse(resolvedPath).root
-      if (!isRoot) {
-        try {
-          await fs.promises.mkdir(downloadPath, { recursive: true })
-        } catch (e) {
-          // Ignore EEXIST; rethrow others
-          if (!(e instanceof Error) || (e as NodeJS.ErrnoException).code !== 'EEXIST') {
-            throw e
-          }
-        }
-      }
-
-      // Start download
-      await new Promise<void>((resolve, reject) => {
-        const startUrl = new URL(downloadUrl)
-        let writer: fs.WriteStream | undefined
-        let filePath: string | undefined
-
-        // Sanitize file names for the current OS and ensure .osz extension
-        const sanitizeFileName = (name: string): string => {
-          const invalidChars = /[<>:\\"/|?*]/g
-          let safe = name.replace(invalidChars, ' ').replace(/\s+/g, ' ').trim()
-          // Disallow trailing periods or spaces on Windows
-          safe = safe.replace(/[ .]+$/g, '')
-          if (!/\.osz$/i.test(safe)) {
-            safe = `${safe}.osz`
-          }
-          return safe
-        }
-
-        const makeRequest = (targetUrl: URL, redirectCount = 0): void => {
-          const protocol = targetUrl.protocol === 'http:' ? http : https
-          const req = protocol.request(
-            targetUrl,
-            {
-              headers: {
-                'User-Agent': 'osu-beatmap-backup/1.0 (+https://github.com)'
-              }
-            },
-            (response) => {
-              // Handle redirects
-              if (
-                response.statusCode &&
-                response.statusCode >= 300 &&
-                response.statusCode < 400 &&
-                response.headers.location
-              ) {
-                if (redirectCount >= 5) {
-                  reject(new Error('Too many redirects'))
-                  return
-                }
-                try {
-                  const nextUrl = new URL(response.headers.location, targetUrl)
-                  req.destroy()
-                  makeRequest(nextUrl, redirectCount + 1)
-                  return
-                } catch {
-                  reject(new Error('Invalid redirect URL'))
-                  return
-                }
-              }
-
-              if (response.statusCode !== 200) {
-                reject(new Error(`Failed to download: ${response.statusCode}`))
-                return
-              }
-
-              const totalSize = parseInt(response.headers['content-length'] || '0', 10)
-
-              // Get filename from Content-Disposition header or fallback to beatmapsetId
-              let fileName = `${task.beatmapsetId}.osz`
-              const contentDisposition = response.headers['content-disposition']
-              if (contentDisposition) {
-                const matches = /filename[^;=\n]*=((['"]).*?\2|[^;\n]*)/.exec(contentDisposition)
-                if (matches && matches[1]) {
-                  fileName = matches[1].replace(/['"]/g, '')
-                }
-              }
-              fileName = sanitizeFileName(fileName)
-              filePath = path.join(downloadPath, fileName)
-
-              // Update task with discovered file name
-              task.fileName = fileName
-              this.emit(DownloadEvent.TASK_UPDATED, task)
-
-              // Create write stream
-              writer = fs.createWriteStream(filePath)
-              let downloadedBytes = 0
-              const startTime = Date.now()
-              let lastUpdate = startTime
-              let lastBytes = 0
-
-              response.on('data', (chunk) => {
-                downloadedBytes += chunk.length
-                const currentTime = Date.now()
-                const timeDiff = (currentTime - lastUpdate) / 1000
-                const bytesDiff = downloadedBytes - lastBytes
-
-                if (timeDiff >= 1) {
-                  task.speed = bytesDiff / timeDiff
-                  task.progress = totalSize ? Math.round((downloadedBytes / totalSize) * 100) : 0
-                  task.remainingTime = totalSize
-                    ? Math.round((totalSize - downloadedBytes) / task.speed)
-                    : 0
-                  this.emit(DownloadEvent.TASK_UPDATED, task)
-                  lastUpdate = currentTime
-                  lastBytes = downloadedBytes
-                }
-              })
-
-              response.on('error', (err) => reject(err))
-              writer.on('error', (err) => reject(err))
-
-              response.pipe(writer)
-
-              writer.on('finish', () => {
-                // Update mirror health on success
-                const mirrorName = task.mirror.name
-                const health = this.mirrorHealth.get(mirrorName) || {
-                  success: 0,
-                  failure: 0,
-                  avgResponseTime: 0
-                }
-                health.success++
-                health.avgResponseTime =
-                  (health.avgResponseTime * (health.success - 1) + (Date.now() - startTime)) /
-                  health.success
-                this.mirrorHealth.set(mirrorName, health)
-
-                task.status = 'completed'
-                task.progress = 100
-                task.speed = 0
-                task.remainingTime = 0
-                task.filePath = filePath
-                this.emit(DownloadEvent.TASK_COMPLETED, task)
-                resolve()
-                // Evaluate queue completion after each task finishes
-                this.checkQueueCompletion()
-              })
-            }
-          )
-
-          req.on('error', (error) => {
-            try {
-              if (writer) {
-                writer.close()
-              }
-            } catch {
-              /* empty */
-            }
-            if (filePath) {
-              fs.unlink(filePath, () => {})
-            }
-            reject(error)
-          })
-
-          // 30s timeout
-          req.setTimeout(30000, () => {
-            req.destroy(new Error('Request timeout'))
-          })
-
-          req.end()
-          // Track request for possible cancellation
-          // eslint-disable-next-line @typescript-eslint/no-explicit-any
-          ;(task as any).request = req
-        }
-
-        makeRequest(startUrl)
+      const { startTime } = await downloadFile(task, taskDownloadPath, (updatedTask) => {
+        this.emit(DownloadEvent.TASK_UPDATED, updatedTask)
       })
+
+      // Update mirror health on success
+      const mirrorName = task.mirror.name
+      const health = this.mirrorHealth.get(mirrorName) || {
+        success: 0,
+        failure: 0,
+        avgResponseTime: 0
+      }
+      health.success++
+      health.avgResponseTime =
+        (health.avgResponseTime * (health.success - 1) + (Date.now() - startTime)) / health.success
+      this.mirrorHealth.set(mirrorName, health)
+
+      this.emit(DownloadEvent.TASK_COMPLETED, task)
+      this.checkQueueCompletion()
     } catch (error) {
       console.error(`Download failed for ${task.beatmapsetId}:`, error)
 
@@ -538,20 +188,22 @@ class DownloadService extends EventEmitter {
       health.failure++
       this.mirrorHealth.set(mirrorName, health)
 
-      // If error is due to abort, don't try next mirror
       if (error instanceof Error && error.message === 'Download aborted') {
         task.status = 'waiting'
         task.error = 'Download cancelled'
         this.emit(DownloadEvent.TASK_UPDATED, task)
+        // Race condition guard: if resumeQueue() ran before this abort settled,
+        // re-add the task now since resumeQueue already iterated past it.
+        if (!this.isPaused && this.currentMirrors.length && this.currentOptions) {
+          this.queue.add(() => this.downloadTask(task, this.currentMirrors, this.currentOptions!))
+        }
         return
       }
 
-      // Try next mirror
       const currentIndex = availableMirrors.indexOf(task.mirror)
       const nextIndex = (currentIndex + 1) % availableMirrors.length
 
       if (nextIndex === 0) {
-        // All mirrors failed, enter cooldown
         task.status = 'error'
         task.error = error instanceof Error ? error.message : 'Download failed'
         this.emit(DownloadEvent.TASK_ERROR, task)
@@ -559,23 +211,21 @@ class DownloadService extends EventEmitter {
         if (!this.cooldownTimeout) {
           this.cooldownTimeout = setTimeout(() => {
             this.cooldownTimeout = undefined
-            // Retry failed tasks
-            for (const [, task] of this.tasks) {
-              if (task.status === 'error') {
-                task.status = 'waiting'
-                task.mirror = availableMirrors[0]
-                this.emit(DownloadEvent.TASK_UPDATED, task)
-                this.queue.add(() => this.downloadTask(task, availableMirrors, options))
+            for (const [, t] of this.tasks) {
+              if (t.status === 'error') {
+                t.status = 'waiting'
+                t.mirror = availableMirrors[0]
+                this.emit(DownloadEvent.TASK_UPDATED, t)
+                this.queue.add(() => this.downloadTask(t, availableMirrors, options))
               }
             }
           }, this.cooldownPeriod)
         }
-        // If we are not retrying immediately, check if queue is now completed
+
         if (this.queue.size === 0 && this.queue.pending === 0) {
           this.checkQueueCompletion()
         }
       } else {
-        // Try next mirror
         task.mirror = availableMirrors[nextIndex]
         task.status = 'waiting'
         this.emit(DownloadEvent.TASK_UPDATED, task)
@@ -588,33 +238,48 @@ class DownloadService extends EventEmitter {
     const { getWaitForDownloadsOnPause } = await import('./settingsStore')
     const shouldWaitForCompletion = getWaitForDownloadsOnPause()
 
-    const activeDownloads = Array.from(this.tasks.values()).filter(
-      (task) => task.status === 'downloading'
-    )
+    this.isPaused = true
+    // Clear queued-but-not-started tasks from PQueue so they don't drain
+    // through downloadTask (hitting the isPaused check and exiting) while paused.
+    // They remain in this.tasks with status 'waiting' and will be re-added on resume.
+    this.queue.clear()
 
-    if (activeDownloads.length > 0 && !shouldWaitForCompletion) {
-      // Cancel active downloads and add them back to queue
-      for (const task of activeDownloads) {
-        if (task.request) {
-          task.request.destroy()
+    if (!shouldWaitForCompletion) {
+      for (const task of this.tasks.values()) {
+        if (task.status === 'downloading') {
+          // Passing the error triggers the 'Download aborted' catch branch in downloadTask
+          task.request?.destroy(new Error('Download aborted'))
         }
       }
     }
 
-    // If shouldWaitForCompletion is true, active downloads will complete naturally
-
-    this.isPaused = true
     this.emit(DownloadEvent.QUEUE_PAUSED)
   }
 
   public resumeQueue(): void {
     this.isPaused = false
+
+    if (this.currentMirrors.length && this.currentOptions) {
+      for (const task of this.tasks.values()) {
+        if (task.status === 'waiting') {
+          this.queue.add(() => this.downloadTask(task, this.currentMirrors, this.currentOptions!))
+        }
+      }
+    }
+
     this.emit(DownloadEvent.QUEUE_RESUMED)
   }
 
   public clearQueue(): void {
+    for (const task of this.tasks.values()) {
+      if (task.status === 'downloading') {
+        task.request?.destroy(new Error('Download aborted'))
+      }
+    }
     this.queue.clear()
     this.tasks.clear()
+    this.currentMirrors = []
+    this.currentOptions = null
     this.emit(DownloadEvent.QUEUE_CLEARED)
   }
 
@@ -630,10 +295,7 @@ class DownloadService extends EventEmitter {
     return this.queue.pending
   }
 
-  public getMirrorHealth(): Map<
-    string,
-    { success: number; failure: number; avgResponseTime: number }
-  > {
+  public getMirrorHealth(): Map<string, MirrorHealth> {
     return this.mirrorHealth
   }
 }
